add model_sql to modelsplitter
file to make changes to specific versions of the schematisation.
format should be json with the first key equal to the
`name` column in model_settings.xlsx. The dict within this schema version
should have one query per key. This query will be applied to the database.
the keyname here is irrelevant.

example:

{
    "0d1d_test": {"query1": "UPDATE v2_orifice SET connection_node_end_id=386 WHERE code LIKE 'KDU-Q-1804'"}
}

will update the connection node of an orifice only for the 0d1d_test model.