Update main.py
update main to select which function to used according to the selection done: It could be: compare DAMO WITH DAMO, and then select specific layer or select them all. It could be also  compare DAMO WITH 3DI, and then select specific layer or select them all. Or it is possible also to select with both.