Create breach_wdepth_damage.py
Hello Wietse, I file where I am creating a set of function that will be helpfull in the future at the moment we want to do the job a bit easier and maybe fasters. It contains the following function:

def clip_DEM: it clip any raster and save it into a vrt raster
def grid_selection: It select the grids that contains water so it expend less time and memory when it has to create the final waterdepth raster
def get_paths:  is a simple function that retrieve the paths that I am going to use to have all the scenarios per regions. If you want to do especfici scenarios it also helps
def calculate_depth_raster. it calculate the depth raster.
def indirect_direct. It calculate the different cost: direct, indirect and total, with the help of the function: calculate_damage_raster
def sum_total_raster: it sum the values per pixel and save it into a csv file with the function save_damage_csv
def create_pgn_damage this function create the bargraph for the final maps.