Update breach_wdepth_damage.py
In this commit I added some comments to clarify function and also to make it more readable.
For differents functions I introduced the EPSG and spatialResolution as a paramenter.
For the Grid Selectiong Function I shorter the way to check if the active_cells where empty or not, and also I skipped one step by deleting  buffer_explode and added into the dissolve.