Update gemaal_intermediate_converter.py
Delete the for loop to add values of waterlevel for summer and for winter, and change it for a shorter way. Remove some not needed columns.
Add new columns to know the sort_peilgebied and de sort_steerfpeil.
Add the black space after Parameters,
Remove not needed imports,
Correct typos