update scirpt that download the data from 3di.  Update paths
It is importnat to notice that the code line
start_time = datetime.strptime(gr.lines.dt_timestamps[0].split(".")[0], "%Y-%m-%dT%H:%M:%S")

might not work. This error has to be with a piece of code line that N&S must fix. I send a note to them so they can fix that.